Create AuthenticatedControllerTest that can be extended in test classes that need a token
