+++ backend/src/test/java/ch/japt/epj/controller/ExerciseControllerTests.java
@@ -4,18 +4,12 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import ch.japt.epj.security.JwtTokenProvider;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
@@ -26,26 +20,9 @@ import org.springframework.transaction.annotation.Transactional;
 @SpringBootTest
 @Transactional
 @AutoConfigureMockMvc
-public class ExerciseControllerTests {
-
-  private String completeToken;
+public class ExerciseControllerTests extends AuthenticatedControllerTest {
 
   @Autowired private MockMvc mvc;
-
-  @Autowired private JwtTokenProvider tokenProvider;
-
-  @Autowired private AuthenticationManager authenticationManager;
-
-  @Before
-  public void getToken() {
-    String validEmail = "jonas.kugler@hsr.ch";
-    String validPassword = "jonas";
-    Authentication authentication =
-        authenticationManager.authenticate(
-            new UsernamePasswordAuthenticationToken(validEmail, validPassword));
-    SecurityContextHolder.getContext().setAuthentication(authentication);
-    completeToken = "Bearer " + tokenProvider.generateToken(authentication);
-  }
 
   @Test
   public void getExerciseSuccess() throws Exception {
